+++ app/src/main/java/im/ene/toro/sample/flexible/FlexiblePlayerViewHolder.java
@@ -45,12 +45,19 @@ public class FlexiblePlayerViewHolder extends RecyclerView.ViewHolder implements
 
   SimpleExoPlayerViewHelper helper;
   Uri mediaUri;
+  View.OnClickListener clickListener;
 
   @BindView(R.id.player) SimpleExoPlayerView playerView;
 
   public FlexiblePlayerViewHolder(View itemView) {
     super(itemView);
     ButterKnife.bind(this, itemView);
+  }
+
+  public void setClickListener(View.OnClickListener clickListener) {
+    this.clickListener = clickListener;
+    itemView.setOnClickListener(clickListener);
+    playerView.setOnClickListener(clickListener);
   }
 
   @NonNull @Override public View getPlayerView() {