+++ toro-core/src/main/java/im/ene/toro/exoplayer/ExoPlayerHelper.java
@@ -18,7 +18,6 @@ package im.ene.toro.exoplayer;
 
 import android.content.Context;
 import android.os.Handler;
-import android.os.Looper;
 import android.support.annotation.FloatRange;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
@@ -27,10 +26,10 @@ import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.DefaultRenderersFactory;
 import com.google.android.exoplayer2.DefaultRenderersFactory.ExtensionRendererMode;
 import com.google.android.exoplayer2.ExoPlaybackException;
-import com.google.android.exoplayer2.ExoPlayer;
 import com.google.android.exoplayer2.ExoPlayerFactory;
 import com.google.android.exoplayer2.ParserException;
 import com.google.android.exoplayer2.PlaybackParameters;
+import com.google.android.exoplayer2.Player;
 import com.google.android.exoplayer2.RenderersFactory;
 import com.google.android.exoplayer2.SimpleExoPlayer;
 import com.google.android.exoplayer2.Timeline;
@@ -77,10 +76,9 @@ public final class ExoPlayerHelper {
 
   private static final String TAG = "ToroLib:ExoPlayer";
 
-  // instance is unchanged, but inner fields are changeable.
-  @NonNull final PlaybackInfo playbackInfo = new PlaybackInfo();
-
-  final Context context;  // will obtain from playerView context, should be Application context.
+  // Instance is unchanged, but inner fields are changeable.
+  final PlaybackInfo playbackInfo = new PlaybackInfo();
+  @NonNull final Context context;  // Application context, will obtain from playerView context.
   @NonNull final SimpleExoPlayerView playerView;
   @ExtensionRendererMode final int extensionMode;
   final Handler mainHandler;
@@ -95,7 +93,7 @@ public final class ExoPlayerHelper {
   boolean shouldAutoPlay;
   boolean needRetrySource;
 
-  ArrayList<ExoPlayer.EventListener> eventListeners;
+  ArrayList<Player.EventListener> eventListeners;
 
   public ExoPlayerHelper(@NonNull SimpleExoPlayerView playerView,
       @ExtensionRendererMode int extensionMode, boolean playWhenReady) {
@@ -103,7 +101,7 @@ public final class ExoPlayerHelper {
     this.context = playerView.getContext().getApplicationContext();
     this.extensionMode = extensionMode;
     this.shouldAutoPlay = playWhenReady;
-    this.mainHandler = new Handler(Looper.myLooper());
+    this.mainHandler = new Handler();
   }
 
   public ExoPlayerHelper(@NonNull SimpleExoPlayerView playerView,
@@ -231,7 +229,7 @@ public final class ExoPlayerHelper {
     return new PlaybackInfo(playbackInfo.getResumeWindow(), playbackInfo.getResumePosition());
   }
 
-  public void addEventListener(@NonNull ExoPlayer.EventListener eventListener) {
+  public void addEventListener(@NonNull Player.EventListener eventListener) {
     if (this.eventListeners == null) {
       this.eventListeners = new ArrayList<>();
     }
@@ -240,7 +238,7 @@ public final class ExoPlayerHelper {
     if (eventListener != null) this.eventListeners.add(eventListener);
   }
 
-  public void removeEventListener(ExoPlayer.EventListener eventListener) {
+  public void removeEventListener(Player.EventListener eventListener) {
     if (this.eventListeners != null && eventListener != null) {
       this.eventListeners.remove(eventListener);
     }
@@ -277,9 +275,13 @@ public final class ExoPlayerHelper {
     playbackInfo.reset();
   }
 
-  private class ComponentListener implements ExoPlayer.EventListener {
+  private class ComponentListener implements Player.EventListener {
 
     ComponentListener() {
+    }
+
+    @Override public void onRepeatModeChanged(int repeatMode) {
+
     }
 
     @Override public void onTimelineChanged(Timeline timeline, Object manifest) {
@@ -325,7 +327,7 @@ public final class ExoPlayerHelper {
     @Override public void onPlayerStateChanged(boolean playWhenReady, int playbackState) {
       boolean screenOn = isPlaying() && (playbackState >= 2 || playbackState <= 3);
       playerView.setKeepScreenOn(screenOn);
-      if (playbackState == ExoPlayer.STATE_ENDED) {
+      if (playbackState == Player.STATE_ENDED) {
         if (player != null) {
           player.setPlayWhenReady(false);
           player.seekTo(0); // reset playback position for current window
@@ -469,11 +471,11 @@ public final class ExoPlayerHelper {
   }
 
   // Adapter for original EventListener
-  public static class EventListener implements ExoPlayer.EventListener {
+  public static class EventListener implements Player.EventListener {
 
-    private ExoPlayer.EventListener delegate;
+    private Player.EventListener delegate;
 
-    public EventListener(ExoPlayer.EventListener delegate) {
+    public EventListener(Player.EventListener delegate) {
       this.delegate = delegate;
     }
 
@@ -481,8 +483,12 @@ public final class ExoPlayerHelper {
       this.delegate = null;
     }
 
-    public void setDelegate(ExoPlayer.EventListener delegate) {
+    public void setDelegate(Player.EventListener delegate) {
       this.delegate = delegate;
+    }
+
+    @Override public void onRepeatModeChanged(int repeatMode) {
+      if (this.delegate != null) this.delegate.onRepeatModeChanged(repeatMode);
     }
 
     @Override public void onTimelineChanged(Timeline timeline, Object manifest) {